+++ mobile-engage/src/main/java/com/emarsys/mobileengage/geofence/model/Geofence.kt
@@ -3,6 +3,6 @@ package com.emarsys.mobileengage.geofence.model
 data class Geofence(val id: String,
                     val lat: Double,
                     val lon: Double,
-                    val radius: Int,
+                    val radius: Double,
                     val waitInterval: Double?,
                     val triggers: List<Trigger>)
